Add and Delete rows in Database
Sometimes cells for "DialogResource" (TypeId = 11) have the TypeId = 5 (String) and sometimes the intended TypeId = 11, same goes for Characters (TypeId = 6).
This sometimes causes program to crash when entering the cell.
This causes the program to crash when entering / editing the cell with wrong TypeId (unless story reloaded)
Presumably this is a bug from Larian Studio in BG3.
Workarounds:
- Use typeId from first row, if database is not empty (implemented)
- Create columns with typeId previously copied from another StorySave (implemented)
- Save and reload the StorySave (implemented)
Might cause problems ingame because it possibly has a different typeId than intended.